Add lampshade to abstention discrepancy
There's a problem with the bye-laws stemming from
Make Abstentions Make Sense Again.
https://docs.google.com/document/d/1R1ARuz-AjBCocGNOla1V0274Vrlu8WQtIFvYzqch9-E/edit?tab=t.0

Actions & Outcomes suggests there should be a new section 5.4.
but because it wasn't included in Bye-Law Amends Proposed section...

For all intents and purposes it does not exist. and what the bye-laws
say were not updated to include section 5.4.

This is the section that would have said:
"Councillors may choose to abstain from voting to maintain quoracy,
however these votes should not be counted
towards the outcome of the decision".

It's not in the bye-laws currently.
https://docs.google.com/document/d/1EuMLjjJZUCbS4saoLJEZ7tfYRWEoDLXgLn-kU4zsa9Y/edit?tab=t.0

This affects every su council vote since. and may mean,
Cuts & Financial Mismanagement of the Uni (amended or otherwise)
strictly speaking, did not pass.